ex_2_42.scm defined empty board and board position representation
in-diagonal? procedure almost done, needs to be fixed

diagonal works, but in doesnt construct the board because it doesnt have to build on top of something

solution found, finished ex_2_42. Now looking for optimizations

changed safe? in order to optimize

refactor

in-diagonal? corrected, there was a much simple way to determine if two positions are in the same diagonal